backend/requirements.txt
openai
flask
python-dotenv
Flask-Cors
